% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospark.R
\name{geospark}
\alias{geospark}
\title{Generates Spatial Sparklines as Simple Feature Collection (sfc)}
\usage{
geospark(coords, my_stats, width, height, mode = "normal")
}
\arguments{
\item{coords}{Points coordinates}

\item{my_stats}{Values associated to points as a column-wise (long) format. For instance, one column = one date}

\item{width}{Total width of the line}

\item{height}{Total height of the line}

\item{mode}{Normal or Log transformation mode}
}
\value{
returns a geometry collection (sfc)
}
\description{
Generates Spatial Sparklines as Simple Feature Collection (sfc)
}
