% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_geospark.R
\name{one_geospark}
\alias{one_geospark}
\title{Generate a Spatial Sparkline as Simple Feature Geometry (sfg)}
\usage{
one_geospark(coords, values, max_value = NULL, width, height, mode = "normal")
}
\arguments{
\item{coords}{Point coordinates}

\item{values}{Values associated to point}

\item{max_value}{Maximum value}

\item{width}{width}

\item{height}{height}

\item{mode}{normal mode or log mode}
}
\value{
returns a geometry (sfg)
}
\description{
Generate a Spatial Sparkline as Simple Feature Geometry (sfg)
}
