% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGeospark.R
\name{singleGeospark}
\alias{singleGeospark}
\title{Generate a Spatial Sparkline as Simple Feature Geometry (sfg)}
\usage{
singleGeospark(
  coords,
  values,
  max_value = NULL,
  width,
  height,
  direction = "ne",
  mode = "normal"
)
}
\arguments{
\item{coords}{Point coordinates}

\item{values}{Values associated to point}

\item{max_value}{Maximum value}

\item{width}{width}

\item{height}{height}

\item{direction}{Direction of lines, upward (north), downward (south), left (west) or right (east). For instance, "ne" will generate upward right lines.}

\item{mode}{normal mode or log mode}
}
\value{
returns a geometry (sfg)
}
\description{
Generate a Spatial Sparkline as Simple Feature Geometry (sfg)
}
