% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGeosparkBM.R
\name{singleGeosparkBM}
\alias{singleGeosparkBM}
\title{Generates averaged Spatial Sparkline as Simple Feature Geometry (sfg)}
\usage{
singleGeosparkBM(
  my_coords,
  my_values,
  max_value = NULL,
  width,
  height,
  direction = "ne",
  n = 7
)
}
\arguments{
\item{my_coords}{Point coordinates}

\item{my_values}{Values associated to point}

\item{max_value}{Maximum value}

\item{width}{Width}

\item{height}{Height}

\item{direction}{Direction of lines, upward (north), downward (south), left (west) or right (east). For instance, "ne" will generate upward right lines.}

\item{n}{Window size (default = 7)}
}
\value{
returns a geometry (sfg)
}
\description{
Generate a Spatial Sparkline as Simple Feature Geometry (sfg) based on a method by Financial Times Data Visualization Designer John Burn-Murdoch
}
