% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospark.R
\name{geospark}
\alias{geospark}
\title{Calculate geographically placed Sparklines basd upon statistical values}
\usage{
geospark(coords, my_stats, width, height, mode = "normal")
}
\arguments{
\item{coords}{Points coordinates}

\item{my_stats}{Values associated to points}

\item{width}{Width}

\item{height}{Height}

\item{mode}{normal mode or log mode}
}
\value{
returns a geometry collection (sfc)
}
\description{
Calculate geographically placed Sparklines basd upon statistical values
}
