% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_geospark.R
\name{one_geospark}
\alias{one_geospark}
\title{Calculate a geographically placed Sparkline}
\usage{
one_geospark(coords, v, max_v = NULL, width, height, mode = "normal")
}
\arguments{
\item{coords}{Point coordinates}

\item{v}{Values associated to point}

\item{max_v}{Maximum value}

\item{width}{width}

\item{height}{height}

\item{mode}{normal mode or log mode}
}
\value{
returns a geometry (sfg)
}
\description{
Calculate a geographically placed Sparkline
}
